Remove EnemyCornerJumperMovement
