--- Assets/Scripts/Enemy/EnemyCornerJumperMovement.cs
@@ -1,21 +1,0 @@
-﻿using UnityEngine;
-
-namespace Enemy
-{
-    public class EnemyCornerJumperMovement : EnemyMovementBase
-    {        
-        [SerializeField] private JumpOverCorner _cornerJump;
-
-        private void Update()
-        {
-            MoveWithCornerHandling();
-        }
-
-        private void MoveWithCornerHandling()
-        {
-            if (_cornerJump.IsCornerReached) return;
-
-            Move();
-        }
-    }
-}
