--- Assets/Scripts/Enemy/EnemyCornerJumperMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 39776509b2236964c8c140076702dce2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
