--- Assets/Scripts/Enemy/EnemyMovementBase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 78881a4f669237c4bad41287db8c1da4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
